+++ lib/src/resource.ts
@@ -264,11 +264,6 @@ export class Resource {
       this.get(properties.commit.lastCommit)?.toString(),
     );
 
-    console.log(
-      'this.commitBuilder.previousComit',
-      this.commitBuilder.previousCommit,
-    );
-
     // Cloning the CommitBuilder to prevent race conditions, and keeping a back-up of current state for when things go wrong during posting.
     const oldCommitBuilder = this.commitBuilder.clone();
     this.commitBuilder = new CommitBuilder(this.getSubject());
@@ -284,7 +279,9 @@ export class Resource {
     const endpoint = new URL(this.getSubject()).origin + `/commit`;
     try {
       this.commitError = null;
-      await postCommit(commit, endpoint);
+      const createdCommit = await postCommit(commit, endpoint);
+      // TODO: Handle `previousCommit mismatch` error, retry if possible
+      this.commitBuilder.setPreviousCommit(createdCommit.id);
       return this.getSubject();
     } catch (e) {
       // If it fails, revert to the old resource with the old CommitBuilder